tests/wpt/web-platform-tests/css/css-flexbox/parsing/flex-direction-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Flexible Box Layout: getComputedStyle().flexDirection</title>
<link rel="help" href="https://drafts.csswg.org/css-flexbox/#propdef-flex-direction">
<meta name="assert" content="flex-direction computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("flex-direction", "row");
test_computed_value("flex-direction", "row-reverse");
test_computed_value("flex-direction", "column");
test_computed_value("flex-direction", "column-reverse");
</script>
</body>
</html>
